+++ CHANGELOG.md
@@ -8,12 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [[Unreleased]]
 ### Added:
 - Function to parse the final account balances from a transaction's metadata
+- Function to parse order book changes from a transaction's metadata
 - Support for Ed25519 seeds that don't use the `sEd` prefix
 
 ### Fixed:
 - Typing for factory classmethods on models
-
-### Fixed:
 - Use properly encoded transactions in `Sign`, `SignFor`, and `SignAndSubmit`
 
 ## [1.6.0] - 2022-06-02